backend/requirements.txt
sqlalchemy
flask-mysql
-U flask-cors
-U Flask-SQLAlchemy
flask-bcrypt
